UI Renders textures and defaults to blank texture

+++ HenkelEngine/res/shaders/UI.vert
@@ -1,14 +1,14 @@
-#version 330 core
+#version 440 core
 layout (location = 0) in vec3 a_Position;
 layout (location = 1) in vec4 a_Color;
 layout (location = 2) in vec2 a_TexCoord;
-layout (location = 3) in int a_TexIndex;
+layout (location = 3) in float a_TexIndex;
 
 uniform mat4 U_ViewProjection;
 uniform mat4 U_Model;
 
 out vec2 v_TexCoord;
-out int v_TexIndex;
+out float v_TexIndex;
 out vec4 v_Color;
 
 void main()
